requirements.txt
networkx==3.4.2
cloudscraper==1.2.71
PyYAML==6.0.2